fix: fixes after rebase on header service refactor, some more docs for rpc_test.go
