+++ service/header/service.go
@@ -59,8 +59,8 @@ func (s *Service) GetByHeight(ctx context.Context, height uint64) (*header.Exten
 }
 
 // Head returns the ExtendedHeader of the chain head.
-func (s *Service) Head(ctx context.Context) (*ExtendedHeader, error) {
-	return s.syncer.store.Head(ctx)
+func (s *Service) Head(ctx context.Context) (*header.ExtendedHeader, error) {
+	return s.store.Head(ctx)
 }
 
 // IsSyncing returns the status of sync
